nuevos cambio

+++ LogicDeNegocio/Requests/EmpleadoRequest.cs
@@ -8,7 +8,7 @@ namespace LogicDeNegocio.Requests
         // Propiedades generadas automáticamente
         public int IdPersona { get; set; }
          public string Empresa { get; set; }
-         public float Sueldo { get; set; }
+         public decimal Sueldo { get; set; }
          public DateTime? FechaContrato { get; set; }
          public int IdCiudad { get; set; }
 }
